+++ examples/cryptocurrency-advanced/backend/src/main.rs
@@ -13,14 +13,14 @@
 // limitations under the License.
 
 use exonum_cli::NodeBuilder;
-use exonum_cryptocurrency_advanced as cryptocurrency;
+use exonum_cryptocurrency_advanced::CryptocurrencyService;
 
 fn main() -> Result<(), failure::Error> {
     exonum::helpers::init_logger().unwrap();
     NodeBuilder::new()
-        .with_service(cryptocurrency::CryptocurrencyService)
+        .with_rust_service(CryptocurrencyService)
         // Starts cryptocurrency instance with given id and name
         // immediately after genesis block creation.
-        .with_default_service(cryptocurrency::CryptocurrencyService)
+        .with_default_rust_service(CryptocurrencyService)
         .run()
 }